requirements.txt
openai==1.93.1
requests==2.32.4
pydantic==2.11.7
graphviz==0.21
IPython==9.4.0
pytz==2025.2
llama-index==0.12.47
llama-index-llms-openai-like==0.4.0
langchain==0.3.26
langchain-openai==0.3.27
langgraph==0.5.1
langgraph-cli[inmem]==0.3.3
debugpy==1.8.14
autogen-agentchat==0.6.2
autogen-ext[openai]==0.6.2
semantic-kernel==1.34.0
pyperclip==1.9.0
python-dotenv==1.1.1
semantic-kernel==1.34.0
fastmcp==2.10.2
fastapi==0.116.0
langchain-mcp-adapters==0.1.8
uvicorn==0.35.0
aiohttp==3.12.13
azure-identity==1.23.0
aiohttp-sse==2.2.0